backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vendor-contract-compliance-analyzer-backend"
version = "0.1.0"
description = "FastAPI backend for vendor contract compliance analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "chromadb>=1.0.15,<2.0.0",
  "fastapi>=0.116.0,<1.0.0",
  "google-genai>=1.9.0,<2.0.0",
  "google-generativeai>=0.8.5,<1.0.0",
  "pypdf>=5.4.0,<6.0.0",
  "python-docx>=1.1.2,<2.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "orjson>=3.10.0,<4.0.0",
  "sqlalchemy>=2.0.38,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
